Add Station class as a property type
Add a new class which extends Space called Station.

This class is similar to Property, but for stations. Hence it leaves out
elements such as houses and the rent is now dependant on how many other
stations the owner owns.

The player now holds differnt "owns" lists for properties, stations and
utilities (not yet added)